+++ src/Microsoft.DotNet.XHarness.CLI/CommandArguments/Arguments/ErrorPatternsFileArgument.cs
@@ -1,11 +1,11 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
 using System;
 using System.IO;
 
-namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.Wasm;
+namespace Microsoft.DotNet.XHarness.CLI.CommandArguments;
 
 internal class ErrorPatternsFileArgument : StringArgument
 {
