+++ src/Microsoft.DotNet.XHarness.CLI/CommandArguments/Arguments/SymbolMapFileArgument.cs
@@ -2,11 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.Wasi;
+namespace Microsoft.DotNet.XHarness.CLI.CommandArguments;
 
 internal class SymbolMapFileArgument : StringArgument
 {
-    public SymbolMapFileArgument() : base("symbol-map=|s=", "Path to wasm symbol map file generated by wasi")
+    public SymbolMapFileArgument() : base("symbol-map=|s=", "Path to wasm symbol map file generated by emscripten/wasi")
     {
     }
 }
