--- src/Microsoft.DotNet.XHarness.CLI/CommandArguments/WASM/Arguments/SymbolMapFileArgument.cs
@@ -1,12 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.Wasm;
-
-internal class SymbolMapFileArgument : StringArgument
-{
-    public SymbolMapFileArgument() : base("symbol-map=|s=", "Path to wasm symbol map file generated by emscripten")
-    {
-    }
-}
