+++ src/Microsoft.DotNet.XHarness.CLI/Commands/WASM/JS/WasmTestCommand.cs
@@ -140,12 +140,12 @@ internal class WasmTestCommand : XHarnessCommand<WasmTestCommandArguments>
             var stdoutFilePath = Path.Combine(Arguments.OutputDirectory, "wasm-console.log");
             File.Delete(stdoutFilePath);
 
-            var symbolicator = WasmSymbolicatorBase.Create(Arguments.SymbolicatorArgument.GetLoadedTypes().FirstOrDefault(),
+            var symbolicator = SymbolicatorBase.Create(Arguments.SymbolicatorArgument.GetLoadedTypes().FirstOrDefault(),
                                                            Arguments.SymbolMapFileArgument,
                                                            Arguments.SymbolicatePatternsFileArgument,
                                                            logger);
 
-            var logProcessor = new WasmTestMessagesProcessor(xmlResultsFilePath,
+            var logProcessor = new TestMessagesProcessor(xmlResultsFilePath,
                                                              stdoutFilePath,
                                                              logger,
                                                              Arguments.ErrorPatternsFile,
