--- src/Microsoft.DotNet.XHarness.CLI/Commands/WASM/WasmLogMessage.cs
@@ -1,7 +1,0 @@
-namespace Microsoft.DotNet.XHarness.CLI.Commands.Wasm;
-
-class WasmLogMessage
-{
-    public string? method { get; set; }
-    public string? payload { get; set; }
-}
